Terminate pool if the timeout is reached
I was under the impression, that trying to `get` an `AsyncResult` with a
timeout not only raises a `TimeoutError` if the worker isn't finished
yet, but also terminates the worker. Apparently this is not the case, so
`terminate()` has to be called on the pool, if we know that some workers
are still active after the timeout is reached.