Make progress logging less verbose
I found that when hitting 100% of the allotted time for the parallel
processes, the progress info gets logged repeatedly. This is probably
because the loop termination condition and the check for when to log
progress overlap precisely at this point. This commits fixes this by
removing the overlapping part from the process logging condition.